HomePage ui component Edit
